src/requirements.txt
npyscreen==4.10.5
PyYAML==5.1.2
SQLAlchemy==1.3.8
